+++ jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   testEnvironment: 'node',
   preset: 'ts-jest',
-  roots: ['<rootDir>/src'],
+  roots: ['<rootDir>/test'],
 }
