+++ test/create.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@jest/globals'
-import { MoveBuilder } from '../build'
+import { MoveBuilder } from '../src/build'
 import path from 'path'
 describe('create and clean new move package', () => {
   const contractDir = path.resolve(__dirname, 'contract/new')
